requirements.txt
discord.py >= 2.0.0
google-genai
python-dotenv
aiohttp